requirements.txt
PyYAML
pandas
certifi==2020.4.5.1
chardet==3.0.4
click==7.1.2
Flask==1.1.2
gunicorn==20.0.4
idna==2.9
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
oauthlib==3.1.0
requests==2.23.0
requests-oauthlib==1.3.0
urllib3==1.25.9
Werkzeug==1.0.1
